refactor(ui): inline useAppRefreshActions into hooks/data.ts and delete
Single-consumer hook (only hooks/data.ts used it after earlier inlining).
The 4 trivial query invalidators and the complex thread-refresh
dedup logic now live directly in useRefresh().
